+++ smok/extras/pred_database/pickling.py
@@ -8,7 +8,7 @@ from .base import BasePredicateDatabase
 class PicklingPredicateDatabase(BasePredicateDatabase):
     def __init__(self, path: str):
         self.__path = path
-        self.__predicates = []
+        self.__predicates = {}
         if os.path.exists(self.__path):
             with open(self.__path, 'rb') as f_in:
                 self.__predicates = pickle.load(f_in)
@@ -18,8 +18,14 @@ class PicklingPredicateDatabase(BasePredicateDatabase):
             pickle.dump(self.__predicates, f_out, pickle.HIGHEST_PROTOCOL)
 
     def get_all_predicates(self) -> tp.List[tp.Dict]:
-        return self.__predicates
+        return list(self.__predicates.values())
+
+    def update_predicate(self, v: tp.Dict) -> None:
+        self.__predicates[v['predicate_id']] = v
+        self.__sync()
 
     def set_new_predicates(self, v: tp.List[tp.Dict]):
-        self.__predicates = v
+        self.__predicates = {}
+        for dct in v:
+            self.__predicates[dct['predicate_id']] = dct
         self.__sync()